implement tests and fix found issues

+++ shared/src/commonMain/kotlin/com/malakiapps/whatsappclone/domain/use_cases/MigrateToGoogleAccountUseCase.kt
@@ -5,14 +5,15 @@ import com.malakiapps.whatsappclone.domain.common.EmailNotFound
 import com.malakiapps.whatsappclone.domain.common.Response
 import com.malakiapps.whatsappclone.domain.common.getOrNull
 import com.malakiapps.whatsappclone.domain.messages.AnonymousUserMessageRepository
+import com.malakiapps.whatsappclone.domain.messages.ConversationBrief
 import com.malakiapps.whatsappclone.domain.messages.MessagesRepository
 import com.malakiapps.whatsappclone.domain.user.ANONYMOUS_EMAIL
 import com.malakiapps.whatsappclone.domain.user.AnonymousUserAccountRepository
 import com.malakiapps.whatsappclone.domain.user.AuthenticatedUserAccountRepository
 import com.malakiapps.whatsappclone.domain.user.None
 import com.malakiapps.whatsappclone.domain.user.Profile
-import com.malakiapps.whatsappclone.domain.user.Some
 import com.malakiapps.whatsappclone.domain.user.SignInResponse
+import com.malakiapps.whatsappclone.domain.user.Some
 import com.malakiapps.whatsappclone.domain.user.UserContactUpdate
 
 
@@ -26,20 +27,29 @@ class MigrateToGoogleAccountUseCase(
         return signInResponse.authenticationContext.email?.let { availableEmail ->
             //MIGRATE THE MESSAGES
             //Get the existing messages
-            anonymousUserMessageRepository.exportAllUserMessages(owner = availableEmail).getOrNull()?.let { existingMessages ->
+            anonymousUserMessageRepository.exportAllUserMessages(owner = ANONYMOUS_EMAIL).getOrNull()?.let { existingMessages ->
                 //Add them to the new account
-                val result = userMessagesRepository.importAllUserMessages(owner = availableEmail, rawConversation = existingMessages)
+                if(existingMessages.messages.isNotEmpty()){
+                    val conversationBrief = ConversationBrief(
+                        newMessageCount = 0,
+                        messageId = existingMessages.messages.first().messageId,
+                        sender = availableEmail,
+                        value = existingMessages.messages.first().value,
+                        sendStatus = existingMessages.messages.first().attributes.sendStatus,
+                        time = existingMessages.messages.first().time
+                    )
+                    val result = userMessagesRepository.importAllUserMessages(owner = availableEmail, conversationBrief = conversationBrief, rawConversation = existingMessages)
 
-                result.getOrNull()?.let {
-                    //DELETE THE ANONYMOUS USER MESSAGES
-                    anonymousUserAccountRepository.deleteAccount(email = ANONYMOUS_EMAIL)
-                    anonymousUserMessageRepository.deleteMessages(owner = ANONYMOUS_EMAIL, messageIds = existingMessages.messages.map { it.messageId })
+                    result.getOrNull()?.let {
+                        //DELETE THE ANONYMOUS USER MESSAGES
+                        anonymousUserMessageRepository.deleteMessages(owner = ANONYMOUS_EMAIL, messageIds = existingMessages.messages.map { it.messageId })
+                    }
                 }
             }
 
             //UPDATE THE NEW ACCOUNT WITH THE ANONYMOUS SETTINGS
             val anonymousUserAccount = anonymousUserAccountRepository.getContact(ANONYMOUS_EMAIL).getOrNull()
-            userAccountRepository.upgradeContactFromAnonymous(
+            val result = userAccountRepository.upgradeContactFromAnonymous(
                 userContactUpdate = UserContactUpdate(
                     email = availableEmail,
                     name = anonymousUserAccount?.name?.let { Some(it) } ?: None,
@@ -47,6 +57,9 @@ class MigrateToGoogleAccountUseCase(
                     image = (anonymousUserAccount?.image ?: signInResponse.initialBase64ProfileImage)?.let { Some(it) } ?: None
                 )
             )
+
+            anonymousUserAccountRepository.deleteAccount(email = ANONYMOUS_EMAIL)
+            result
         } ?: Response.Failure(EmailNotFound)
     }
 }
